refactor(TagForm): extract to seperate file
