+++ modules/TicketMetadata.js
@@ -4,7 +4,8 @@ import PropTypes from 'prop-types'
 import {FormGroup, FormControl, Button} from 'react-bootstrap'
 import AV from 'leancloud-storage/live-query'
 
-import {getCustomerServices, UserLabel, getCategoryPathName, CategoriesSelect, depthFirstSearchFind, TagForm} from './common'
+import {getCustomerServices, UserLabel, getCategoryPathName, CategoriesSelect, depthFirstSearchFind} from './common'
+import TagForm from './TagForm'
 import css from './Ticket.css'
 import csCss from './CustomerServiceTickets.css'
 import translate from './i18n/translate'
